Moving helper objects
Moving the helper objects to their own directory and namespace.
